Feat: 추가 포스트 개수 카운팅 api

+++ src/main/java/mangpo/server/controller/PostController.java
@@ -111,7 +111,7 @@ public class PostController {
 
     @GetMapping("/total-count")
     public Result<?> getTotalCount() {
-        Long totalCount = postService.findTotalCount();
+        Integer totalCount = postService.findTotalCount();
         return new Result<>(totalCount);
     }
